% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cor_ts_name.R
\name{cor_ts_name}
\alias{cor_ts_name}
\title{Correctify Training Set Name}
\usage{
cor_ts_name(ts_name, ts_list)
}
\arguments{
\item{ts_name}{Training Set's name to correct}

\item{ts_list}{Vector listing the existing Training Sets}
}
\value{
Correct Training Set's name or NULL
}
\description{
Check if the name is already used, or null, and if not, take off the special chars from the name. If yes, return NULL.
}
\examples{
ts_list_ex <- c("test1","test2")
ts_name1 <- "test1"
ts_name2 <- "test3"
ts_name3 <- "test_4"
cor_ts_name(ts_name1,ts_list_ex)
cor_ts_name(ts_name2,ts_list_ex)
cor_ts_name(ts_name3,ts_list_ex)
}
