% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_samples.R
\name{find_samples}
\alias{find_samples}
\title{Samples Folders}
\usage{
find_samples(sample_folder_content)
}
\arguments{
\item{sample_folder_content}{Vector of names of a folder's content}
}
\value{
The names of the samples' folders from a vector of names
}
\description{
Find the names of Samples Folders in a given path
}
\examples{
sample_folder_content <- list.files(system.file("example","Samples", package = "ZooImageUI")) 
# Attention qu'ici, il vaudrait mieux utiliser un dossier du package, en 
# utilisant des données échantillon du package
find_samples(sample_folder_content)
}
