% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_is_script_correct.R
\name{is_script_good_model}
\alias{is_script_good_model}
\title{Is Script Good Model ?}
\usage{
is_script_good_model(models_folder_path, selected_script)
}
\arguments{
\item{models_folder_path}{Path to scripts/models.}

\item{selected_script}{Selected script.}
}
\value{
TRUE if the script is good or FALSE if it isn't. Plus a message explaining what is the problem or not.
}
\description{
Try different parameters of a script to check if it is correct to make a model.
}
\examples{
folder_path <- system.file("example", package = "zooimageui")
is_correct <- is_script_good_model(folder_path, "ex1_mlRforest.R")

}
