% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_is_script_res_correct.R
\name{is_script_good_results}
\alias{is_script_good_results}
\title{Is Script Good Results ?}
\usage{
is_script_good_results(results_folder_path, selected_script)
}
\arguments{
\item{results_folder_path}{Path to scripts for calculations.}

\item{selected_script}{Selected script.}
}
\value{
TRUE if the script is good or FALSE if it isn't. Plus a message explaining what is the problem or not.
}
\description{
Try different parameters of a script to check if it is correct to make results.
}
\examples{
folder_path <- system.file("example", package = "ZooImageUI")
is_correct <- is_script_good_results(folder_path, "ex1_zoo_abundance.R")

}
