% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_res_dl.R
\name{res_dl}
\alias{res_dl}
\title{Zip the Results}
\usage{
res_dl(data_folder_path, name, res, file)
}
\arguments{
\item{data_folder_path}{Path to data folder.}

\item{name}{Name of the files for downloading.}

\item{res}{Results to zip.}

\item{file}{Name of the zip file.}
}
\value{
A zip file containing the .txt and .RData files.
}
\description{
Zip Results of {zooimageui} in .txt and .RData files.
}
\examples{
# To Do...
}
