% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_res_save.R
\name{res_save}
\alias{res_save}
\title{Save the Results}
\usage{
res_save(data_folder_path, name, res)
}
\arguments{
\item{data_folder_path}{Folder containing all the data for {zooimageui}.}

\item{name}{Name of the files for saving.}

\item{res}{Results to save.}
}
\value{
\code{TRUE} if the save worked, \code{FALSE} if it didn't. Plus a message if there is an error.
}
\description{
Save Results of {zooimageui} in .txt and .RData files.
}
\examples{
# To Do...
}
