% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\title{Samples Folders}
\usage{
samples(Sample_folder_files)
}
\arguments{
\item{Sample_folder_files}{}
}
\value{
The names of the samples' sub-folders in a main "Sample" folder
}
\description{
Samples Folders
}
\examples{
Sample_folder_files <- list.files("~/shared/data/zooimage_data/Sample")
samples(Sample_folder_files)
}
