% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_upload_ts.R
\name{upload_ts}
\alias{upload_ts}
\title{Upload TS}
\usage{
upload_ts(upload_input, ts_folder_path, existing_ts)
}
\arguments{
\item{upload_input}{Shiny input file}

\item{ts_folder_path}{Path to folder to put ts in}

\item{existing_ts}{List of the existing training sets}
}
\value{
TRUE if it worked, FALSE if an error occurred
}
\description{
Get the zipped training set from upload, unzip it, unlink the zip file.
}
\examples{
# Need a reactive context
}
