% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooimageui-package.R
\docType{package}
\name{zooimageui-package}
\alias{zooimageui-package}
\title{zooimageui}
\description{
A web app to make an interface for the {zooimage} package.
}
\section{Application Shiny}{

\itemize{
\item To launch the Shiny application : use \code{\link[=run_app]{run_app()}}
}
}

