% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\title{Samples Folders}
\usage{
samples(Sample_folder_content)
}
\arguments{
\item{Sample_folder_content}{Vector of names of a folder's content}
}
\value{
The names of the samples' folders from a vector of names
}
\description{
Description
}
\examples{
Sample_folder_content <- list.files("~/shared/data/zooimage_data/Sample") 
# Attention qu'ici, il vaudrait mieux utiliser un dossier du package, en 
# utilisant des données échantillon du package
samples(Sample_folder_content)
}
