% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZooImageUI-package.R
\docType{package}
\name{ZooImageUI-package}
\alias{ZooImageUI-package}
\title{ZooImageUI}
\description{
It is a web app to make an interface to the program ZooImage.
}
\section{Application Shiny}{


- To launch the Shiny application : use [run_app()]
}

